Flatten domain CLI commands
