+++ iruka/commands.go
@@ -9,7 +9,8 @@ var Commands = []cli.Command{
 	cmdCreate,
 	cmdConfig,
 	cmdDestroy,
-	cmdDomain,
+	cmdDomains,
+	cmdDomainAdd,
 	cmdPs,
 	cmdDeploy,
 	cmdOpen,
